Allow multiple directories in DbLedgerStorage
